Rename model files(Python naming convention)
